Progress updating tests for new create-app, update-app refactor
