+++ test/src/tasks/start/backend_old/permissions.test.ts
@@ -1,0 +1,86 @@
+// import { assert } from 'chai'
+// import { createDao } from '~/src/tasks/start/backend/create-dao'
+// import { deployDaoFactory } from '~/src/tasks/start/backend/bases/deploy-dao-factory'
+// import { readArapp, getAppId } from '~/src/utils/arappUtils'
+// import { deployImplementation } from '~/src/tasks/start/backend/app'
+// import { createProxy } from '~/src/tasks/start/backend/proxy'
+// import { AragonAppJson } from '~/src/types'
+// import { useDefaultEnvironment } from '~/test/test-helpers/useEnvironment'
+// import {
+//   setAllPermissionsOpenly,
+//   ANY_ADDRESS
+// } from '~/src/tasks/start/backend/set-permissions'
+// import {
+//   KernelInstance,
+//   ACLContract,
+//   ACLInstance,
+//   DAOFactoryInstance
+// } from '~/typechain'
+
+// describe.skip('permissions.ts', function() {
+//   useDefaultEnvironment()
+
+//   let dao: KernelInstance
+//   let acl: ACLInstance
+//   let arapp: AragonAppJson
+//   let app: any
+
+//   before('set up dao with app', async function() {
+//     const daoFactory: DAOFactoryInstance = await deployDaoFactory(
+//       this.env.artifacts
+//     )
+//     dao = await createDao(this.env.web3, this.env.artifacts, daoFactory.address)
+
+//     const ACL: ACLContract = this.env.artifacts.require('ACL')
+//     acl = await ACL.at(await dao.acl())
+
+//     const implementation = await deployImplementation(this.env.artifacts)
+//     const appId = getAppId('counter')
+//     app = await createProxy(
+//       implementation,
+//       appId,
+//       dao,
+//       this.env.web3,
+//       this.env.artifacts,
+//       []
+//     )
+
+//     arapp = readArapp()
+//   })
+
+//   describe('when setAllPermissionsOpenly is called', function() {
+//     before('call setAllPermissionsOpenly', async function() {
+//       await setAllPermissionsOpenly(
+//         dao,
+//         app,
+//         arapp,
+//         this.env.web3,
+//         this.env.artifacts
+//       )
+//     })
+
+//     it('properly sets the INCREMENT_ROLE permission', async function() {
+//       assert.equal(
+//         await acl.hasPermission(
+//           ANY_ADDRESS,
+//           app.address,
+//           await app.INCREMENT_ROLE()
+//         ),
+//         true,
+//         'Invalid permission.'
+//       )
+//     })
+
+//     it('properly sets the DECREMENT_ROLE permission', async function() {
+//       assert.equal(
+//         await acl.hasPermission(
+//           ANY_ADDRESS,
+//           app.address,
+//           await app.DECREMENT_ROLE()
+//         ),
+//         true,
+//         'Invalid permission.'
+//       )
+//     })
+//   })
+// })
