+++ test/src/tasks/start/backend_old/proxy.test.ts
@@ -1,0 +1,90 @@
+// import { assert } from 'chai'
+// import { useDefaultEnvironment } from '~/test/test-helpers/useEnvironment'
+// import { deployImplementation } from '~/src/tasks/start/backend/deploy-implementation'
+// import { createProxy, updateProxy } from '~/src/tasks/start/backend/proxy'
+// import { readArapp, getAppEnsName, getAppId } from '~/src/utils/arappUtils'
+// import { setAllPermissionsOpenly } from '~/src/tasks/start/backend/set-permissions'
+// import { createDao } from '~/src/tasks/start/backend/create-dao'
+// import { deployDaoFactory } from '~/src/tasks/start/backend/bases/deploy-dao-factory'
+
+// describe.skip('proxy.ts', function() {
+//   let dao, implementation, proxy, appId
+
+//   describe('when in the counter project', async function() {
+//     useDefaultEnvironment()
+
+//     before('create a dao and an app implementation', async function() {
+//       ;({ dao, implementation } = await _createDaoAndAppImplementation(
+//         this.env.web3,
+//         this.env.artifacts
+//       ))
+//     })
+
+//     describe('when creating a counter proxy', async function() {
+//       before('create a counter proxy and set permissions', async function() {
+//         appId = getAppId(getAppEnsName())
+
+//         proxy = await createProxy(
+//           implementation,
+//           appId,
+//           dao,
+//           this.env.web3,
+//           this.env.artifacts,
+//           []
+//         )
+
+//         const arapp = readArapp()
+//         await setAllPermissionsOpenly(
+//           dao,
+//           proxy,
+//           arapp,
+//           this.env.web3,
+//           this.env.artifacts
+//         )
+//       })
+
+//       describe('when updating the counter proxy', async function() {
+//         let newImplementation
+
+//         before('deploy new implementation', async function() {
+//           const CounterV1 = this.env.artifacts.require('CounterV1')
+
+//           newImplementation = await CounterV1.new()
+//         })
+
+//         before('update the proxy', async function() {
+//           await updateProxy(newImplementation, appId, dao, this.env.web3)
+//         })
+
+//         it('the dao references the correct implementation for it', async function() {
+//           assert.equal(
+//             newImplementation.address,
+//             await dao.getApp(await dao.APP_BASES_NAMESPACE(), appId),
+//             'Incorrect implementation in proxy'
+//           )
+//         })
+
+//         it('reports the correct hardcoded version', async function() {
+//           const version = (await proxy.getVersion()).toString()
+
+//           assert.equal(version, '1', 'Incorrect counter proxy version')
+//         })
+
+//         itBehavesLikeACounterContract()
+//       })
+//     })
+//   })
+// })
+
+// async function _createDaoAndAppImplementation(
+//   web3,
+//   artifacts
+// ): Promise<{ dao: any; implementation: any }> {
+//   const daoFactory = await deployDaoFactory(artifacts)
+
+//   const dao = await createDao(web3, artifacts, daoFactory.address)
+
+//   const implementation = await deployImplementation(artifacts)
+
+//   return { dao, implementation }
+// }
